requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0,<0.10.20
customtkinter>=5.2.0
pystray>=0.19.0
Pillow>=10.0.0
playsound==1.2.2
